Refactor Makefile so that Device/Bluetooth Configurators work
The main Makefile is used to parse the project configuration, so
splitting up the Makefile into multiple files broke BSP configuration.
Given this, I had to switch to a hackier solution to display the splash.
Luckily, the prebuild target existed in start.mk, so I could reuse that
to modify the generated source code and rename the BLE server.